// src/types/content.ts
export interface ContentData {
  landingPage: {
    title: string;
    paragraphs: {
      greeting: string;
      professional: string;
      passions: string;
      philosophy: string;
    };
    links: {
      instagram: string;
      pinterest: string;
      youtube: string;
    };
    footer: string;
  };
}